Change backend to sqlite
Use the Sequelize orm to handle database access. This changes the
backend to sqlite, but any supported database could be used with
configuration to be implemented in a later commit.

This change also considerably refactors the model by consolidating the
db, reader, and saver files into the main model file Record.

Primary keys have been changed to an autoincrementing integer for the
convenience of users.

This change contains breaking changes and will not be compatible with
current installs of the project.